+++ app.js
@@ -4,6 +4,9 @@ const path = require('path');
 const express = require('express');
 const bodyParser = require('body-parser');
 const session = require('express-session')
+const csurf = require('csurf');
+const cookieParser = require('cookie-parser');
+const helmet = require('helmet');
 
 const gameRoutes = require('./routes/game');
 const errorController = require('./controllers/error');
@@ -14,16 +17,25 @@ app.set('view engine', 'ejs');
 app.set('views', 'views');
 app.set('trust proxy', 1);
 
+app.use(helmet({
+    contentSecurityPolicy: false
+}));
+
 app.use(express.static(path.join(__dirname, 'public')));
 
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(bodyParser.json());
+
 app.use(session({
     secret: 'onenation',
-    cookie: { maxAge: 60000 },
+    cookie: { maxAge: 60000, sameSite: 'strict'},
     resave: true,
     saveUninitialized: false,
 }));
+
+app.use(cookieParser());
+app.use(csurf({ cookie: false }));
+
 
 app.use('/', gameRoutes);
