+++ package.json
@@ -27,9 +27,12 @@
     "mocha": "^8.2.1"
   },
   "dependencies": {
+    "cookie-parser": "^1.4.5",
+    "csurf": "^1.11.0",
     "dotenv": "^8.2.0",
     "ejs": "^2.7.4",
     "express": "^4.17.1",
-    "express-session": "^1.17.1"
+    "express-session": "^1.17.1",
+    "helmet": "^4.2.0"
   }
 }
