+++ views/game/game.ejs
@@ -31,6 +31,10 @@
     <br />
 </div>
 
+<script type="text/javascript">
+    var _csrf = "<%=csrfToken%>";
+</script>
+<script type="text/javascript" src="js/clue.js"></script>
 <script type="text/javascript" src="js/game.js"></script>
 
 <%- include('../includes/footer.ejs') %>
